+++ rtthost/src/main.rs
@@ -1,10 +1,11 @@
 use probe_rs::{config::TargetSelector, DebugProbeInfo, Probe};
 use probe_rs_rtt::{Channels, Rtt, RttChannel, ScanRegion};
+
+use clap::Parser;
 use std::io::prelude::*;
 use std::io::{stdin, stdout};
 use std::sync::mpsc::{channel, Receiver};
 use std::thread;
-use structopt::StructOpt;
 
 #[derive(Debug, PartialEq, Eq)]
 enum ProbeInfo {
@@ -26,7 +27,9 @@ impl std::str::FromStr for ProbeInfo {
     }
 }
 
-fn parse_scan_region(mut src: &str) -> Result<ScanRegion, Box<dyn std::error::Error>> {
+fn parse_scan_region(
+    mut src: &str,
+) -> Result<ScanRegion, Box<dyn std::error::Error + Send + Sync + 'static>> {
     src = src.trim();
     if src.is_empty() {
         return Ok(ScanRegion::Ram);
@@ -50,13 +53,13 @@ fn parse_scan_region(mut src: &str) -> Result<ScanRegion, Box<dyn std::error::Er
     }
 }
 
-#[derive(Debug, StructOpt)]
-#[structopt(
+#[derive(Debug, clap::Parser)]
+#[clap(
     name = "rtthost",
     about = "Host program for debugging microcontrollers using the RTT (real-time transfer) protocol."
 )]
 struct Opts {
-    #[structopt(
+    #[clap(
         short,
         long,
         default_value = "0",
@@ -64,31 +67,31 @@ struct Opts {
     )]
     probe: ProbeInfo,
 
-    #[structopt(
+    #[clap(
         short,
         long,
         help = "Target chip type. Leave unspecified to auto-detect."
     )]
     chip: Option<String>,
 
-    #[structopt(short, long, help = "List RTT channels and exit.")]
+    #[clap(short, long, help = "List RTT channels and exit.")]
     list: bool,
 
-    #[structopt(
+    #[clap(
         short,
         long,
         help = "Number of up channel to output. Defaults to 0 if it exists."
     )]
     up: Option<usize>,
 
-    #[structopt(
+    #[clap(
         short,
         long,
         help = "Number of down channel for keyboard input. Defaults to 0 if it exists."
     )]
     down: Option<usize>,
 
-    #[structopt(
+    #[clap(
         long,
         default_value="",
         parse(try_from_str=parse_scan_region),
@@ -103,7 +106,7 @@ fn main() {
 }
 
 fn run() -> i32 {
-    let opts = Opts::from_args();
+    let opts = Opts::parse();
 
     let probes = Probe::list_all();
 
@@ -289,7 +292,7 @@ fn list_channels(channels: &Channels<impl RttChannel>) {
         println!(
             "  {}: {} (buffer size {})",
             chan.number(),
-            chan.name().as_deref().unwrap_or("(no name)"),
+            chan.name().unwrap_or("(no name)"),
             chan.buffer_size(),
         );
     }